BACKEND/requirements.txt
Flask
flask_mongoengine
flask_jwt_extended
python-dotenv